+++ packages/chat/src/components/NewFooter/NewFooter.css.ts
@@ -1,19 +1,29 @@
 import { keyframes, style } from '@vanilla-extract/css';
 import { recipe } from '@vanilla-extract/recipes';
 
+import { timingFunction } from '@/styles/animations';
 import { COLORS } from '@/styles/colors';
 import { PALETTE } from '@/styles/colors.css';
 import { FAMILY } from '@/styles/font';
 import { SIZES } from '@/styles/sizes';
 import { transition } from '@/styles/transitions';
-import { widgetContainer } from '@/views/ChatWidget/styles.css';
+import { chatIsClosed, chatIsOpen } from '@/views/ChatWidget/styles.css';
 
 const BUTTON_ROW_MARGIN = 10;
+
+const fadeIn = keyframes({
+  from: {
+    opacity: 0,
+  },
+  to: {
+    opacity: 1,
+  },
+});
 
 const fadeInAndSlideUp = keyframes({
   from: {
     opacity: 0,
-    transform: 'translateY(10px)',
+    transform: 'translateY(40px)',
   },
   to: {
     opacity: 1,
@@ -28,34 +38,19 @@ const fadeOutSlideDown = keyframes({
   },
   to: {
     opacity: 0,
-    transform: 'translateY(10px)',
-  },
-});
-
-const fadeIn = keyframes({
-  from: {
-    opacity: 0,
-  },
-  to: {
-    opacity: 1,
+    transform: 'translateY(40px)',
   },
 });
 
 export const footerContainer = style({
   width: '100%',
-
   selectors: {
-    [`.${widgetContainer.classNames.variants.withChat.true} &`]: {
-      animation: `${fadeInAndSlideUp} 300ms ease-in`,
-      opacity: 1,
-      pointerEvents: 'auto',
-      transform: 'translateY(0)',
+    [`.${chatIsOpen} &`]: {
+      animation: `${fadeInAndSlideUp} .6s ${timingFunction.gentle} forwards`,
     },
-    [`.${widgetContainer.classNames.variants.withChat.false} &`]: {
-      animation: `${fadeOutSlideDown} 300ms ease-out`,
-      opacity: 0,
-      pointerEvents: 'none',
-      transform: 'translateY(100%)',
+    [`.${chatIsClosed} &`]: {
+      animationDelay: '0.5s', // Added delay
+      animation: `${fadeOutSlideDown} 300ms ${timingFunction.gentle} forwards`,
     },
   },
 });